Add health check endpoints
DistroBuildSync can now respond to HTTP health check requests from
OpenShift. The default HTTP port is 8080, but can be specified with the
-p/--port argument.

This implements a startup check at `https://<server>:<port>/startup` and
a liveness check at `https://<server>:<port>/alive`.

While adding this, I realized that the error handling for load_config()
was inconsistent (and would crash in some places), so I cleaned that up
as well.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>