Fix errback for failed tasks
We have an extra errback routine for failed tasks to check whether the
EBS timeout was reached. It's intended to cancel the Koji task in that
case, however errbacks are not meant to be asynchronous generator
functions. This patch modifies the cancel_timed_out_task() routine to
return immediately and enqueue the cancelation to occur out-of-band in
the mainloop.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>