Define TMPDIR
The on-the-fly updating of distrobuildsync-config is handled by a git
checkout in a location created by tempfile.TemporaryDirectory, which
defaults to /tmp but is overrided by TMPDIR, TEMP, or TMP envvars.  As we
are no longer relying on /tmp being writable (per the previous commit),
this too needs to be set in run.sh.  Also, using a single variable
throughout makes it easier to readjust in the future.