Make prepare_side_tag() wait for initial tagging
Currently, it just "fires-and-forgets" the initial tags, relying on the
fact that we usually wait for the buildroot to regenerate. This actually
could have been a race-condition, but we tended to be okay since tagging
is usually very fast.

However, we need to ensure we wait for tagging to complete before we
submit a newly-created side-tag as a Bodhi update, because here we were
hitting a race-condition almost every time.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>